v1.0

+++ controller/courseController.php
@@ -36,14 +36,27 @@
     function editModule() {
       if (isset($_POST['name']) && isset($_POST['header_image']) && isset($_POST['preview']) && isset($_POST['content'])){
         if (!empty($_POST['name']) && !empty($_POST['header_image']) && !empty($_POST['preview']) && !empty($_POST['content'])) {
-          if ($this->courseManagement->createModule()) {
-            echo json_encode(array('success' => 1));
+          if (file_exists($_POST['header_image'])) {
+            $contenidoBinario = file_get_contents($_POST['header_image']);
+            $imagenComoBase64 = base64_encode($contenidoBinario);
+            $_POST['header_image'] = $imagenComoBase64;
+            if ($this->courseManagement->createModule()) {
+              echo json_encode(array('success' => 1));
+            }
+            else {
+              echo json_encode(
+                array(
+                  'success' => 0, 
+                  'msg'     => 'No se ha podido crear el módulo.'
+                )
+              );
+            }
           }
           else {
             echo json_encode(
               array(
                 'success' => 0, 
-                'msg'     => 'El correo o la contraseña no coinciden.'
+                'msg'     => 'La imagen de cabecera no existe.'
               )
             );
           }
